[mono] Move culture-info-tables.h to a separate directory
We'd like to remove it from the dotnet/runtime copy of mono since we use ICU
there for globalization like the rest of .NET 5.

We move the file to a separate directory and then don't sync it between
mono/mono and dotnet/runtime.  The dotnet/runtime copy will be empty
