+++ mono/culture/.gitignore
@@ -1,0 +1,12 @@
+/Makefile
+/Makefile.in
+/.libs
+/.deps
+/*.lo
+/*.la
+/*.o
+/*.a
+/semantic.cache
+/.project
+/.cproject
+/TAGS
