+++ mono/culture/Makefile.am
@@ -1,0 +1,6 @@
+MAKEFLAGS := $(MAKEFLAGS) --no-builtin-rules
+
+# there's nothing to compile in this directory, so we use EXTRA_DIST
+# if we built a convenience library here, we wouldn't need this
+EXTRA_DIST = culture-info-tables.h
+
